fix duplicates from icloud
