+++ src/components/ActivityToThread.tsx
@@ -42,7 +42,6 @@ const ActivitytoThread = (props: any) => {
             onClick={() => {
               setShowDesc(false);
               setSeeThreadAssign(false);
-              // closePopover();
               dispatch({
                 type: 'ADD_ACTIVITY_TO_THREAD',
                 activity,
